1. add script tranlate form 2. fix bugs for SqlAnalyser: (1) sqlserver: insert clause no column (2) add handle for: delete from join clause (3) oracle: update set join clause 3. fix bugs for dbconverter: quotation char incorrect, such as some search condition and alias

+++ DatabaseConverter/SqlAnalyser.Core/StatementScriptBuilder/TSqlStatementScriptBuilder.cs
@@ -141,7 +141,18 @@ namespace SqlAnalyser.Core
             }
             else if (statement is DeleteStatement delete)
             {
-                this.AppendLine($"DELETE FROM {delete.TableName}");
+                bool hasJoin = AnalyserHelper.IsFromItemsHaveJoin(delete.FromItems);
+
+                if (!hasJoin)
+                {
+                    this.AppendLine($"DELETE FROM {delete.TableName}");
+                }
+                else
+                {
+                    this.AppendLine($"DELETE {delete.TableName}");
+
+                    this.BuildFromItems(delete.FromItems);
+                }
 
                 if (delete.Condition != null)
                 {
